Small adjustments in fan-out consumer
