+++ clientlibrary/worker/fan-out-shard-consumer.go
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 VMware, Inc.
+ * Copyright (c) 2021 VMware, Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction, including
